ai_research_framework/requirements.txt
# AI Research Framework — Python Dependencies
# Install: pip install -r ai_research_framework/requirements.txt
# All packages must be documented in docs/context.md after installation.

# ── Orchestration ─────────────────────────────────────────────────────────────
langgraph>=0.2.0
pydantic-ai>=0.0.14

# ── LLM API ───────────────────────────────────────────────────────────────────
anthropic>=0.40.0

# ── Forecasting models ────────────────────────────────────────────────────────
pmdarima>=2.0.4             # AutoARIMA
prophet>=1.1.5              # Prophet (Meta)
lightgbm>=4.3.0             # LightGBM
xgboost>=2.0.3              # XGBoost
scikit-learn>=1.4.0         # Ridge + calibration utilities

# ── Database connectivity ─────────────────────────────────────────────────────
pyodbc>=5.1.0               # ODBC Driver 18 for SQL Server (must be installed separately)
azure-identity>=1.16.0      # Azure AD service principal auth (ClientSecretCredential)
python-dotenv>=1.0.0        # Load .env into os.environ

# ── Data manipulation ─────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0             # Efficient CSV/parquet loading

# ── Memory profiling ──────────────────────────────────────────────────────────
# tracemalloc: Python standard library (no install needed)
memory-profiler>=0.61.0     # Supplementary line-by-line profiling

# ── Statistical evaluation ────────────────────────────────────────────────────
statsmodels>=0.14.0         # Diebold-Mariano test
scipy>=1.12.0               # Isotonic regression (calibration)
